feat: todo 변경 사항 적용
