--- week07/김시온/todo copy/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Todo List</title>
-  </head>
-  <body class="m-0 flex place-items-center min-w-[320px] min-h-screen bg-white text-[#213547] dark:bg-[#242424] dark:text-white/90">
-    <div id="root" class="max-w-screen-xl mx-auto p-8 text-center h-[590px] flex flex-col justify-start bg-[#2a2a2a] rounded-[12px]"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
-</html>
