--- week07/김시온/todo copy/src/index.css
@@ -1,11 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-@layer components {
-  .scrollbar-none::-webkit-scrollbar {
-    display: none;
-  }
-  .scrollbar-none {
-    scrollbar-width: none;
-  }
-}
