--- week07/김시온/todo copy/tailwind.config.js
@@ -1,9 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  darkMode: 'media',
-  content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}', './node_modules/flowbite/**/*.js'],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
