+++ week07/김시온/todo/src/components/TodoItem.jsx
@@ -1,71 +1,88 @@
-import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrash, faPenToSquare, faFloppyDisk, faTimes } from '@fortawesome/free-solid-svg-icons';
+import { useState } from 'react';
 
-function TodoItem({ item, deleteTodo, toggleComplete, updateTodoContent }) {
+function TodoItem({ todo, updateTodo, deleteTodo, toggleComplete }) {
   const [isEditing, setIsEditing] = useState(false);
-  const [editText, setEditText] = useState(item.content);
+  const [content, setContent] = useState(todo.content);
 
-  const handleEdit = () => {
-    setIsEditing(true);
-    setEditText(item.content);
-  };
+  function handleChange(e) {
+    setContent(e.target.value);
+  }
 
-  const handleCancel = () => {
-    setIsEditing(false);
-  };
-
-  const handleSave = () => {
-    if (editText.trim() === '') {
-      alert('내용을 입력해주세요.');
-      return;
-    }
-    updateTodoContent(item.id, editText);
-    setIsEditing(false); // 저장 후 보기 모드로 전환
-  };
-
-  const handleInputChange = (e) => {
-    setEditText(e.target.value);
-  };
-
-  const handleKeyDown = (e) => {
-    if (e.key === 'Enter') {
+  function handleKeyDown(e) {
+    if (e.key == 'Enter') {
       handleSave();
-    } else if (e.key === 'Escape') {
+    } else if (e.key == 'Escape') {
       handleCancel();
     }
-  };
+  }
+
+  async function handleSave() {
+    if (content.trim() === '') {
+      alert('내용을 입력해주세요');
+      return;
+    }
+    updateTodo(todo.id, content);
+    setIsEditing(false);
+  }
+
+  function handleCancel() {
+    setIsEditing(false);
+  }
+
+  function handleEdit() {
+    setIsEditing(true);
+    setContent(todo.content);
+  }
 
   return (
-    <li className={`bg-[#333] mt-[10px] p-[10px] rounded-lg flex justify-between items-center text-white list ${item.isComplete ? 'line-through opacity-[0.6]' : ''}`}>
-      <input type='checkbox' className='mr-[10px] flex-shrink-0' checked={item.isComplete} onChange={() => toggleComplete(item.id)} disabled={isEditing} />
+    <li className={`bg-[#333] mt-[10px] p-[10px] rounded-lg flex justify-between items-center text-white list ${todo.isComplete ? 'line-through text-opacity-60 ' : ''}`}>
+      <input type='checkbox' className='mr-[10px]' checked={todo.isComplete} onChange={() => toggleComplete(todo.id, todo.isComplete)} disabled={isEditing} />
 
       {isEditing ? (
-        <input type='text' value={editText} onChange={handleInputChange} onKeyDown={handleKeyDown} className='flex-grow bg-[#444] text-white p-1 rounded border border-gray-500 mr-2' autoFocus />
+        <input
+          type='text'
+          value={content}
+          onChange={handleChange}
+          onKeyDown={handleKeyDown}
+          className='dark:bg-[#444] border dark:border-white rounded justify-between mr-1 w-40 h-8 pl-1'
+          autoFocus></input>
       ) : (
-        <span className='flex-grow break-all mr-2' onClick={() => !isEditing && toggleComplete(item.id)} style={{ cursor: isEditing ? 'default' : 'pointer' }}>
-          {item.content}
+        <span className={`${isEditing ? 'cursor-default' : 'cursor-pointer'}`} onClick={() => !isEditing && toggleComplete(todo.id, todo.isComplete)}>
+          {todo.content}
         </span>
       )}
 
-      <div className='flex-shrink-0'>
-        {' '}
-        {/* 버튼 그룹을 div로 묶음 */}
+      <div>
         {isEditing ? (
           <>
-            <button title='저장' className='bg-green-500 text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:bg-green-600 mr-1' onClick={handleSave}>
+            {' '}
+            <button
+              className='bg-[#f9f9f9] dark:bg-green-500 dark:text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:outline hover:bg-green-600 mr-1 hover:outline-2 hover:outline-white '
+              onClick={() => handleSave()}>
               <FontAwesomeIcon icon={faFloppyDisk} />
             </button>
-            <button title='취소' className='bg-gray-500 text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:bg-gray-600' onClick={handleCancel}>
+            <button
+              className='bg-[#f9f9f9] dark:bg-red-500 dark:text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:outline hover:bg-red-600 hover:outline-2 hover:outline-white '
+              onClick={() => handleCancel(todo.id)}>
               <FontAwesomeIcon icon={faTimes} />
             </button>
           </>
         ) : (
           <>
-            <button title='수정' className='bg-blue-500 text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:bg-blue-600 mr-1' onClick={handleEdit}>
+            {' '}
+            <button
+              className={`bg-[#f9f9f9] dark:bg-blue-500 dark:text-white border-none px-[8px] py-[6px] rounded-[5px] mr-1 ${
+                !todo.isComplete ? 'cursor-pointer hover:outline hover:bg-blue-600 hover:outline-2 hover:outline-white' : ''
+              } ${todo.isComplete ? ' opacity-60' : ''}`}
+              onClick={() => handleEdit()}
+              disabled={todo.isComplete}>
               <FontAwesomeIcon icon={faPenToSquare} />
             </button>
-            <button title='삭제' className='bg-red-500 text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer hover:bg-red-600' onClick={() => deleteTodo(item.id)}>
+            <button
+              className='bg-[#f9f9f9] dark:bg-stone-900 dark:text-white border-none px-[8px] py-[6px] rounded-[5px] cursor-pointer  hover:bg-stone-950 hover:outline hover:outline-2 hover:outline-white '
+              onClick={() => deleteTodo(todo.id)}>
               <FontAwesomeIcon icon={faTrash} />
             </button>
           </>
