+++ week07/김시온/todo/src/components/TodoList.jsx
@@ -1,11 +1,11 @@
 import TodoItem from './TodoItem';
 
-function TodoList({ todoList, deleteTodo, toggleComplete, updateTodoContent }) {
+function TodoList({ todoList, updateTodo, deleteTodo, toggleComplete }) {
   return (
     <div id='todoListContainer' className='mt-[10px] max-h-[370px] overflow-y-scroll scrollbar-none'>
       <ul className='p-0'>
-        {todoList.map((item) => (
-          <TodoItem key={item.id} item={item} deleteTodo={deleteTodo} toggleComplete={toggleComplete} updateTodoContent={updateTodoContent} />
+        {todoList.map((todo) => (
+          <TodoItem key={todo.id} todo={todo} updateTodo={updateTodo} deleteTodo={deleteTodo} toggleComplete={toggleComplete} />
         ))}
       </ul>
     </div>
